refactor(frontend): standalone user badge component
